add useEmojiAutocompletion hooks
